frontend/requirements.txt
streamlit
requests
Pillow
opencv-python
